131-aliyun-sendsms/pyproject.toml
[tool.poetry]
name = "aliyun-sendsms"
version = "0.1.0"
description = ""
authors = ["阏男秀 <ynanxiu@qq.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^2.2"
alibabacloud-dysmsapi20170525 = "^2.0.15"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
